fix: MemberService UserService 통합, 기타 기능 추가
